added line follower app
added line follower app

+++ bot/src/drivers/oled_extensions.h
@@ -7,7 +7,7 @@
 void OLED_putFloatVariable(PmodOLED *InstancePtr, int xch, int ych, char * head,
         float value, char * tail);
 
-void OLED_putIntVariable(PmodOLED *InstancePtr, int16_t value);
+void OLED_putIntVariable(PmodOLED *InstancePtr, int32_t value);
 
 void  OLED_RecedeCursor(PmodOLED *InstancePtr);
